+++ __tests__/client/interface.test.tsx
@@ -1,26 +1,51 @@
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
-import Home from '../../src/client/scripts/pages/Home';
-import About from '../../src/client/scripts/pages/About';
-import NotFound from '../../src/client/scripts/pages/NotFound';
+
+import { StaticRouter } from 'react-router-dom';
+
+import App from '../../src/client/scripts/App';
 
 test('Home', () => {
     const tree = renderer
-        .create(<Home />)
+        .create(
+            <StaticRouter location={"/home"}>
+                <App />
+            </StaticRouter>
+        )
         .toJSON();
     expect(tree).toMatchSnapshot();
 });
 
 test('About', () => {
     const tree = renderer
-        .create(<About />)
+        .create(
+            <StaticRouter location={"/about"}>
+                <App />
+            </StaticRouter>
+        )
         .toJSON();
     expect(tree).toMatchSnapshot();
 });
 
 test('NotFound', () => {
     const tree = renderer
-        .create(<NotFound />)
+        .create(
+            <StaticRouter location={`/${randomString(10)}`}>
+                <App />
+            </StaticRouter>
+        )
         .toJSON();
     expect(tree).toMatchSnapshot();
 });
+
+function randomString(length: number) {
+    let text = "";
+    const possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+
+    let i;
+    for (i = 0; i < length; i++) {
+        text += possible.charAt(Math.floor(Math.random() * possible.length));
+    }
+
+    return text;
+}